requirements.txt
flask
flask_socketio
python-dotenv
groq
